Create Jupyter Class and Object Error
Jupyter will give an error when running this code that says: Vehicle() takes no arguments. This happens when I try to create a new object, in this case car1. There doesn't appear to be any errors in the code. It also reads the methods I defined as attributes, and I'm not sure why it does this.